Add files via upload
The files listed in here are small personal projects for financial calculations, which i put together to practice writing code in python. If you have suggestions or feedback feel free to submit them to me, but avoid sending specific code as writing this is more for my own education than for a marketable application.